Export MockConnection
Prior to this commit, projects that used this library we're unable to
take advantage of the mocking infrastructure that exists! This lead to
challenges when trying to test functions that used an
*invdapi.Connection for instance.

This commit remedies that challenge by revising the unexported `mockConnection`
function to an exported `MockConnection`. It also adds some
documentation to help anyone interested in _how_ to use the mocked
connection and how to access it in their own project. Further, it
revises all the calls that existed to the unexported function to use the
exported name.

Thanks for taking a look!

Closes #7