Add Plan field to SubscriptionAddon struct
The Invoiced API docs detail a 'plan' field in the Subscription
AddOn Object but, at the time of this commit, this library does not
include it in its own SubscriptionAddon struct. This lead to
difficulty using plans with subscriptions addons.

After this commit, there will be a 'Plan' field exposed so that users
can specify an addon's plan.