Revise SubscriptionAddon's CatalogItem field
Prior to this commit, attempting to use a CatalogItem as a
SubscriptionAddon resulted in this error:

```
json: cannot unmarshal bool into Go struct field SubscriptionAddon.id of type int64
```

This lead to what seemed like an inability to actually use catalog items
as subscription addons.

This commit revises the CatalogItem field of the SubscriptionAddon
structure. It has been converted from it's own struct to a string.

While the catalog item field within a subscription addon appears to be
undocumented, the documentation around catalog items in the API suggests
that strings are expected.

The revisions in this commit have allowed me to use catalog items as
subscription addons as I'd expected.

Thanks for taking a look, and feel free to let me know if I was just
using it all wrong!