Export makeEndPointURL function
Prior to this commit, the method used to generate an
appropriate endpoint URL (dev mode sandbox, or prod) for calls to the
Invoiced REST API was not exposed to applications using this library.

Because there are cases where consumers of this libary need to use the
REST API directly (in order to get a plan by ID for instance), the
application is required to track the appropriate endpoint URLs
themselves.

This commit is composed of a change to make the makeEndPointURL visible
to the library consumer (as MakeEndPointURL) and propagates that change
to the places in this code base where the method was used.

Closes #11