Add ShipTo field to Subscription
Prior to this commit, if a user wanted to add a shipping address to a
subscription, they were required to make a PATCH or POST request
direclty to the Invoiced REST API.

This can cause issues because creating and updating subscriptions with shipping
addresses requires logic for that field alone to be separate from the
rest of the subscription handling, i.e. more code to maintain and more
places for things to fail.

Thanks for taking a look!