Mitigate error unmarshalling Tax.id
We've run into cases where our request for a subscription
(Subscription.Retrieve) failed with the following message:

`json: cannot unmarshal string into Go struct field Tax.id of type
int64`

Looking into the docs on json.Marshal here
(https://golang.org/pkg/encoding/json/#Marshal) it says that:

> The "string" option signals that a field is stored as JSON inside a JSON-encoded string. It applies only to fields of string, floating point, integer, or boolean types. This extra level of encoding is sometimes used when communicating with JavaScript programs

My hypothesis is that _this_ int64 is different than the others becasue
of how it's used with Avalara? Or another JavaScript program?

Up for suggestions here if there's a preference for an alternative
solution but it seemed to have done the trick for us on our fork.